Do not close settings during account sign in.
First of all, there is no need, since we do not automatically continue after sign in anymore.
Moreover, it causes some trouble with the addon being started in another process when the current listing is updated just in between closing the keyboard for username and opening it again for password entry. This update causes a lot of trouble, like a useless attempt to intialise My List and creating an account session object before new account data is available. The latter is a problem, because this process is used in subsequent invocations of the addon.

Signed-off-by: dimkroon <111366411+dimkroon@users.noreply.github.com>